Split reusable  code
* keep reusable functions to suite
* move test to separate file
